requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-aiplatform
python-dotenv
